pyproject.toml
[project]
name = "pi-random-qr-print"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.1.0",
    "pycups>=2.0.4",
    "qrcode[pil]>=8.0",
    "rpi-gpio>=0.7.1",
]
